Add files via upload
Theres nothing much here right now. Whats written in class1 is just a way to read the client id and client secret. I suppose it works fine, i dont know how to check it. I didnt include the file containing the info. So on your side, create a file where the first line is just the client id and the second line is the client secret and name it "clients.txt"